Update _site.yml
links reference in _site.yml is incorrect for Chronic Wasting Disease tutorial. It looks like they have updated it the tutorial along with the path. As a result the site no longer builds with the correct URLs for that tutorial. I have updated the href text but I can't build the site (incorrect permissions)